+++ client/components/DoughnutChart.jsx
@@ -14,7 +14,7 @@ class DoughnutChart extends React.Component {
           label: 'Errors',
           backgroundColor: ['#BDD358', '#FFFFFF', '#1BB075', '#999799'],
           borderColor: 'transparent',
-          data: [0, 0, 0, 0],
+          data: [25, 25, 25, 25],
         },
       ],
     };
@@ -24,72 +24,72 @@ class DoughnutChart extends React.Component {
     const socket = io.connect('http://localhost:3333');
     // Connects 404 Error Consumer
     socket.on('404_ERRORS_PER_MIN', (data) => {
-      // parse incoming data 
-      const message = JSON.parse(data);
-      
-      // store the current state array in a variable
-      const dataArray = this.state.datasets[0].data;
+      // // parse incoming data 
+      // const message = JSON.parse(data);
 
-      // create a new data variable, spread current array into new array
-      const updatedArray = [...dataArray];
+      // // store the current state array in a variable
+      // const dataArray = this.state.datasets[0].data;
 
-      // push incoming data to new data variable
-      //   console.log("these are updated datasets", updatedDataSets);
-      //   updatedDataSets[0].data.splice(0, 1, message.ERROR_404_COUNT);
-      //   console.log("spliced array:", updatedDataSets.data);
-      updatedArray[0] = message.COUNT;
-      console.log('doughtnut data num:', updatedArray[0]);
+      // // create a new data variable, spread current array into new array
+      // const updatedArray = [...dataArray];
 
-      // set the state with the updated variable
-      //   this.setState({ datasets: updatedDataSets });
-      //   console.log("state", this.state);
+      // // push incoming data to new data variable
+      // //   console.log("these are updated datasets", updatedDataSets);
+      // //   updatedDataSets[0].data.splice(0, 1, message.ERROR_404_COUNT);
+      // //   console.log("spliced array:", updatedDataSets.data);
+      // updatedArray[0] = message.COUNT;
+      // console.log('doughtnut data num:', updatedArray[0]);
+
+      // // set the state with the updated variable
+      // //   this.setState({ datasets: updatedDataSets });
+      // //   console.log("state", this.state);
     });
 
     // Connects 405 Error Consumer
     socket.on('405_ERRORS_PER_MIN', (data) => {
-      const message = JSON.parse(data);
-      const currDataSets = this.state.datasets;
-      // console.log(currDataSets);
+      // const message = JSON.parse(data);
+      // const currDataSets = this.state.datasets;
+      // // console.log(currDataSets);
 
-      // create a new data variable, spread current array into new array
-      const updatedDataSets = [...currDataSets];
+      // // create a new data variable, spread current array into new array
+      // const updatedDataSets = [...currDataSets];
 
-      // push incoming data to new data variable
-      updatedDataSets[0].data.splice(1, 1, message.COUNT);
+      // // push incoming data to new data variable
+      // updatedDataSets[0].data.splice(1, 1, message.COUNT);
 
-      // set the state with the updated variable
-      this.setState({ datasets: updatedDataSets });
+      // // set the state with the updated variable
+      // this.setState({ datasets: updatedDataSets });
     });
 
     // Connects 406 Error Consumer
     socket.on('406_ERRORS_PER_MIN', (data) => {
-      const message = JSON.parse(data);
-      const currDataSets = this.state.datasets;
-      // console.log(currDataSets);
+      // const message = JSON.parse(data);
+      // const currDataSets = this.state.datasets;
+      // // console.log(currDataSets);
 
-      // create a new data variable, spread current array into new array
-      const updatedDataSets = [...currDataSets];
+      // // create a new data variable, spread current array into new array
+      // const updatedDataSets = [...currDataSets];
 
-      // push incoming data to new data variable
-      updatedDataSets[0].data.splice(2, 1, message.COUNT);
+      // // push incoming data to new data variable
+      // updatedDataSets[0].data.splice(2, 1, message.COUNT);
 
-      // set the state with the updated variable
-      this.setState({ datasets: updatedDataSets });
+      // // set the state with the updated variable
+      // this.setState({ datasets: updatedDataSets });
     });
 
     // Connects 407 Error Consumer
     socket.on('407_ERRORS_PER_MIN', (data) => {
-      const message = JSON.parse(data);
-      const currDataSets = this.state.datasets;
-      // console.log(currDataSets);
+      // const message = JSON.parse(data);
+      // const currDataSets = this.state.datasets;
+      // // console.log(currDataSets);
 
-      // create a new data variable, spread current array into new array
-      const updatedDataSets = [...currDataSets];
+      // // create a new data variable, spread current array into new array
+      // const updatedDataSets = [...currDataSets];
 
-      // push incoming data to new data variable
-      updatedDataSets[0].data.splice(3, 1, message.COUNT);
+      // // push incoming data to new data variable
+      // updatedDataSets[0].data.splice(3, 1, message.COUNT);
 
-      this.setState({ datasets: updatedDataSets });
+      // this.setState({ datasets: updatedDataSets });
     });
   }
